+++ package.json
@@ -9,6 +9,9 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@react-native-community/masked-view": "^0.1.11",
+    "@react-navigation/bottom-tabs": "^6.5.11",
+    "@react-navigation/drawer": "^6.6.6",
     "@react-navigation/native": "^6.1.9",
     "@react-navigation/stack": "^6.3.20",
     "expo": "~49.0.13",
@@ -18,8 +21,10 @@
     "react": "18.2.0",
     "react-native": "0.72.5",
     "react-native-animatable": "^1.3.3",
+    "react-native-gesture-handler": "^2.13.4",
     "react-native-modal-dropdown": "^1.0.2",
-    "react-native-safe-area-context": "4.6.3",
+    "react-native-reanimated": "^3.5.4",
+    "react-native-safe-area-context": "^4.6.3",
     "react-native-screens": "~3.22.0",
     "react-native-vector-icons": "^10.0.2"
   },
